UPSTREAM: <carry>: Makefile&co. changes for monitoring container
Because of plugin image and one of downstream commits, the commit
dcbb75e is handicapped and Makefile changes are lost. Also changes to
hack/render-manifests.go are lost.

The same applies to the commit 5476488.

Then hack/ocp-update-bundle-manifests.sh needs to be aware of the
kube-rbac-proxy so we are adding it too.

Then, Makefile needs to have ability to separately configure repo,
namespace, image and tag for kube-rbac-proxy container image. It also
needs ability to configure the namespace in which cluster monitoring is
deployed.

Lastly, hack/ocp-bundle-common.sh needs ability to set kube-rbac-proxy
image name and tag.